requirements.txt
# LLM - DeepSeek (兼容 OpenAI SDK)
openai>=1.12.0

# STT - Whisper 语音识别
faster-whisper>=1.0.0

# TTS - Edge-TTS (微软免费语音合成，无需GPU)
edge-tts>=6.1.9

# Web 服务器
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# 音频处理
numpy>=1.24.0
soundfile>=0.12.1
pydub>=0.25.1

# 工具
python-dotenv>=1.0.0
aiofiles>=23.2.1
